// apps/api/src/db/generated/hire_me/Admin.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { ColumnType, Selectable, Insertable, Updateable } from 'kysely';

/** Identifier type for hire_me.admin */
export type AdminId = number & { __brand: 'AdminId' };

/** Represents the table hire_me.admin */
export default interface AdminTable {
  id: ColumnType<AdminId, never, never>;

  email: ColumnType<string, string, string>;

  password_hash: ColumnType<string, string, string>;
}

export type Admin = Selectable<AdminTable>;

export type NewAdmin = Insertable<AdminTable>;

export type AdminUpdate = Updateable<AdminTable>;