// apps/api/src/db/generated/hire_me/Salary.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { ColumnType, Insertable, Selectable, Updateable } from "kysely";
import { Range } from "postgres-range";
import type { RoleId } from "./Role";
import type { default as SalaryPeriod } from "./SalaryPeriod";

/** Identifier type for hire_me.salary */
export type SalaryId = number & { __brand: "SalaryId" };

/** Represents the table hire_me.salary */
export default interface SalaryTable {
	id: ColumnType<SalaryId, never, never>;

	role_id: ColumnType<RoleId, RoleId, RoleId>;

	salary_range: ColumnType<Range<number>, Range<number>, Range<number>>;

	salary_includes_super: ColumnType<boolean, boolean, boolean>;

	salary_period: ColumnType<SalaryPeriod, SalaryPeriod, SalaryPeriod>;

	salary_currency: ColumnType<string, string, string>;
}

export type Salary = Selectable<SalaryTable>;

export type NewSalary = Insertable<SalaryTable>;

export type SalaryUpdate = Updateable<SalaryTable>;